Refactor Dashboard and database code for consistency.
Reorder elements in Dashboard charts, database migrations, and Nav items for better structure and clarity. Updated database table IDs to use BIGINT for improved scalability. Adjusted UI components, including NavBar logo alignment.

+++ dist-electron/main.js
@@ -309,7 +309,7 @@ const DATABASE_PATH = "./local.db";
 const DB_OPTIONS = {};
 const CREATE_CREDITS_TABLE = `
     CREATE TABLE IF NOT EXISTS credits (
-                                           id INTEGER PRIMARY KEY AUTOINCREMENT,
+                                           id BIGINT PRIMARY KEY AUTOINCREMENT,
                                            date TEXT,
                                            title TEXT,
                                            amount REAL,
@@ -318,7 +318,7 @@ const CREATE_CREDITS_TABLE = `
 `;
 const CREATE_DEBITS_TABLE = `
     CREATE TABLE IF NOT EXISTS debits (
-                                          id INTEGER PRIMARY KEY AUTOINCREMENT,
+                                          id BIGINT PRIMARY KEY AUTOINCREMENT,
                                           date TEXT,
                                           title TEXT,
                                           amount REAL,
